proper use of api key newtype
